backend/requirements.txt
uvicorn
fastapi
pillow
tensorflow
python-multipart